feat: SPA
