+++ components/FlappingBirds.vue
@@ -48,6 +48,8 @@
 </template>
 
 <script>
+import registerAFrameComponent from '../lib/registerAFrameComponent'
+
 const flappingBirdsComponent = {
   init() {
     for (let i = 0; i < 5; i++) {
@@ -74,44 +76,9 @@ const flappingBirdsComponent = {
 }
 export default {
   name: 'FlappingBirds',
-  head() {
-    return {
-      // Scripts required by the 8th Wall Framework
-      script: [
-        // slightly modified version of A-Frame, which fixes some polish concerns
-        { src: '//cdn.8thwall.com/web/aframe/8frame-1.1.0.min.js' },
-        {
-          src: '//cdn.8thwall.com/web/aframe/aframe-physics-system-4.0.1.min.js',
-        },
-        // XR Extras - provides utilities like load screen, almost there, and error handling.
-        // See github.com/8thwall/web/tree/master/xrextras
-        { src: '//cdn.8thwall.com/web/xrextras/xrextras.js' },
-        // Add extras from aframe for animations
-        { src: '//cdn.8thwall.com/web/aframe/aframe-extras-6.1.1.min.js' },
-        // 8thWall Web - Replace the app key here with your own app key (only works on authorised domains)
-        {
-          src: '//apps.8thwall.com/xrweb?appKey=zl9iYLs0UnM13G8kugSsXRboJtbC2OJOZWmGeV4dvmWAKMaq1kwdRIa4PTdy4WvWyR05BG',
-        },
-      ],
-    }
-  },
+
   mounted() {
-    this.on8thWallReady()
-  },
-  methods: {
-    on8thWallReady() {
-      AFRAME.registerComponent('flapping-birds', flappingBirdsComponent)
-    },
+    registerAFrameComponent('flapping-birds', flappingBirdsComponent)
   },
 }
 </script>
-
-<style lang="scss">
-a-scene {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-}
-</style>
