+++ components/TapToPlace.vue
@@ -4,7 +4,6 @@
     <a-scene
       v-pre
       tap-place
-      landing-page
       xrextras-loading
       xrextras-runtime-error
       renderer="colorManagement:true"
@@ -72,6 +71,8 @@
 </template>
 
 <script>
+import registerAFrameComponent from '../lib/registerAFrameComponent'
+
 const tapPlaceComponent = {
   schema: {
     min: { default: 6 },
@@ -115,55 +116,8 @@ const tapPlaceComponent = {
 export default {
   name: 'TapToPlace',
 
-  head() {
-    return {
-      // Scripts required by the 8th Wall Framework
-      script: [
-        // slightly modified version of A-Frame, which fixes some polish concerns
-        { src: '//cdn.8thwall.com/web/aframe/8frame-1.1.0.min.js' },
-        {
-          src: '//cdn.8thwall.com/web/aframe/aframe-physics-system-4.0.1.min.js',
-        },
-
-        // XR Extras - provides utilities like load screen, almost there, and error handling.
-        // See github.com/8thwall/web/tree/master/xrextras
-        { src: '//cdn.8thwall.com/web/xrextras/xrextras.js' },
-        // Add extras from aframe for animations
-        { src: '//cdn.8thwall.com/web/aframe/aframe-extras-6.1.1.min.js' },
-
-        // 8thWall Web - Replace the app key here with your own app key (only works on authorised domains)
-        {
-          src: '//apps.8thwall.com/xrweb?appKey=zl9iYLs0UnM13G8kugSsXRboJtbC2OJOZWmGeV4dvmWAKMaq1kwdRIa4PTdy4WvWyR05BG',
-        },
-      ],
-    }
-  },
   mounted() {
-    this.on8thWallReady()
-  },
-  methods: {
-    on8thWallReady() {
-      AFRAME.registerComponent('tap-place', tapPlaceComponent)
-    },
+    registerAFrameComponent('tap-place', tapPlaceComponent)
   },
 }
 </script>
-
-<style lang="scss">
-a-scene {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-}
-
-#backButton {
-  position: absolute;
-  z-index: 5;
-  left: 1vh;
-  top: 1vh;
-  max-width: 12vw;
-  height: 50px;
-}
-</style>
