+++ lib/eventBus.js
@@ -1,0 +1,6 @@
+import Vue from 'vue'
+
+// as per https://blog.logrocket.com/using-event-bus-in-vue-js-to-pass-data-between-components/
+
+const eventBus = new Vue()
+export default eventBus
