+++ lib/registerAFrameComponent.js
@@ -1,0 +1,5 @@
+export default function (componentName, component) {
+  if (!AFRAME.components[componentName]) {
+    AFRAME.registerComponent(componentName, component)
+  }
+}
